backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
gql==3.5.2
graphql-core==3.2.4
h11==0.16.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
httpx